backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-multipart==0.0.9
pydantic==2.8.2
PyPDF2==3.0.1
python-dotenv==1.0.1
httpx==0.27.2
google-generativeai==0.7.2
pdfminer.six==20240706
pytest==9.0.1
